Move relation code to separate file
